Move to_yaml/from_yaml logic to PipelineConfiguration
